Move unit-test specific hook helpers to own class
To prevent false-positive coverage
