+++ test/unit/lib/Dequeuer.test.js
@@ -11,7 +11,7 @@ const BlinkRetryError = require('../../../src/errors/BlinkRetryError');
 const Dequeuer = require('../../../src/lib/Dequeuer');
 const RetryManager = require('../../../src/lib/RetryManager');
 const FreeFormMessage = require('../../../src/messages/FreeFormMessage');
-const HooksHelper = require('../../helpers/HooksHelper');
+const UnitHooksHelper = require('../../helpers/UnitHooksHelper');
 const MessageFactoryHelper = require('../../helpers/MessageFactoryHelper');
 
 // ------- Init ----------------------------------------------------------------
@@ -20,8 +20,8 @@ chai.should();
 chai.use(sinonChai);
 
 // Setup blink app for each test.
-test.beforeEach(HooksHelper.createRandomQueueInMemory);
-test.afterEach.always(HooksHelper.destroyRandomQueueInMemory);
+test.beforeEach(UnitHooksHelper.createRandomQueueInMemory);
+test.afterEach.always(UnitHooksHelper.destroyRandomQueueInMemory);
 
 // ------- Tests ---------------------------------------------------------------
