+++ test/unit/lib/RetryManager.test.js
@@ -10,7 +10,7 @@ const sinonChai = require('sinon-chai');
 const BlinkRetryError = require('../../../src/errors/BlinkRetryError');
 const RetryManager = require('../../../src/lib/RetryManager');
 const DelayLogic = require('../../../src/lib/DelayLogic');
-const HooksHelper = require('../../helpers/HooksHelper');
+const UnitHooksHelper = require('../../helpers/UnitHooksHelper');
 const MessageFactoryHelper = require('../../helpers/MessageFactoryHelper');
 
 // ------- Init ----------------------------------------------------------------
@@ -19,8 +19,8 @@ chai.should();
 chai.use(sinonChai);
 
 // Setup blink app for each test.
-test.beforeEach(HooksHelper.createRandomQueueInMemory);
-test.afterEach.always(HooksHelper.destroyRandomQueueInMemory);
+test.beforeEach(UnitHooksHelper.createRandomQueueInMemory);
+test.afterEach.always(UnitHooksHelper.destroyRandomQueueInMemory);
 
 // ------- Tests ---------------------------------------------------------------
